+++ src/index.ts
@@ -2,6 +2,6 @@ export { PageFlip } from './PageFlip';
 export { FlipCorner, FlipDirection, FlippingState } from './Flip/Flip';
 export { PageDensity, PageOrientation } from './Page/Page';
 export { Orientation } from './Render/Render';
-export { CoverDensity, ReadingDirection, SizeType } from './Settings';
+export { CoverDensity, DisplayMode, ReadingDirection, SizeType } from './Settings';
 export type { DataType, EventCallback, WidgetEvent } from './Event/EventObject';
 export type { FlipSetting } from './Settings';